+++ R/write.R
@@ -40,6 +40,8 @@
 #'   "double", "backslash" or "none". You can also use `FALSE`, which is
 #'   equivalent to "none". The default is to double the quotes, which is the
 #'   format excel expects.
+#' @param eol The end of line character to use. Most commonly either "\n" for
+#'   Unix style newlines, or "\r\n" for Windows style newlines.
 #' @return `write_*()` returns the input `x` invisibly.
 #' @references Florian Loitsch, Printing Floating-Point Numbers Quickly and
 #' Accurately with Integers, PLDI '10,
@@ -60,13 +62,13 @@
 #'
 #' \dontshow{setwd(.old_wd)}
 write_delim <- function(x, path, delim = " ", na = "NA", append = FALSE,
-                        col_names = !append, quote_escape = "double") {
+                        col_names = !append, quote_escape = "double", eol = "\n") {
   stopifnot(is.data.frame(x))
 
   x_out <- x
   x[] <- lapply(names(x), function(i) output_column(x[[i]], i))
   stream_delim(x, path, delim = delim, col_names = col_names, append = append,
-    na = na, quote_escape = quote_escape)
+    na = na, quote_escape = quote_escape, eol = eol)
 
   invisible(x_out)
 }
@@ -74,19 +76,19 @@ write_delim <- function(x, path, delim = " ", na = "NA", append = FALSE,
 #' @rdname write_delim
 #' @export
 write_csv <- function(x, path, na = "NA", append = FALSE, col_names = !append,
-                      quote_escape = "double") {
+                      quote_escape = "double", eol = "\n") {
   write_delim(x, path, delim = ",", na = na, append = append,
-    col_names = col_names, quote_escape = quote_escape)
+    col_names = col_names, quote_escape = quote_escape, eol = eol)
 }
 
 #' @rdname write_delim
 #' @export
 write_csv2 <- function(x, path, na = "NA", append = FALSE, col_names = !append,
-                       quote_escape = "double") {
+                       quote_escape = "double", eol = "\n") {
   x_out <- x
   x <- change_decimal_separator(x, decimal_mark = ",")
   write_delim(x, path, delim = ";", na = na, append = append,
-    col_names = col_names, quote_escape = quote_escape)
+    col_names = col_names, quote_escape = quote_escape, eol = eol)
   
   invisible(x_out)
 }
@@ -94,7 +96,7 @@ write_csv2 <- function(x, path, na = "NA", append = FALSE, col_names = !append,
 #' @rdname write_delim
 #' @export
 write_excel_csv <- function(x, path, na = "NA", append = FALSE,
-                            col_names = !append, delim = ",", quote_escape = "double") {
+                            col_names = !append, delim = ",", quote_escape = "double", eol = "\n") {
 
   stopifnot(is.data.frame(x))
 
@@ -104,7 +106,7 @@ write_excel_csv <- function(x, path, na = "NA", append = FALSE,
 
   x[] <- lapply(x, output_column)
   stream_delim(x, path, delim, col_names = col_names, append = append,
-    na = na, bom = !append, quote_escape = quote_escape)
+    na = na, bom = !append, quote_escape = quote_escape, eol = eol)
 
   invisible(x_out)
 }
@@ -112,7 +114,7 @@ write_excel_csv <- function(x, path, na = "NA", append = FALSE,
 #' @rdname write_delim
 #' @export
 write_excel_csv2 <- function(x, path, na = "NA", append = FALSE,
-                             col_names = !append, delim = ";", quote_escape = "double") {
+                             col_names = !append, delim = ";", quote_escape = "double", eol = "\n") {
   x_out <- x
   x <- change_decimal_separator(x, decimal_mark = ",")
 
@@ -120,15 +122,15 @@ write_excel_csv2 <- function(x, path, na = "NA", append = FALSE,
   x[datetime_cols] <- lapply(x[datetime_cols], format, "%Y/%m/%d %H:%M:%S")
 
   x[] <- lapply(x, output_column)
-  write_excel_csv(x, path, na, append, col_names, delim, quote_escape = quote_escape)
+  write_excel_csv(x, path, na, append, col_names, delim, quote_escape = quote_escape, eol = eol)
 
   invisible(x_out)
 }
 
 #' @rdname write_delim
 #' @export
-write_tsv <- function(x, path, na = "NA", append = FALSE, col_names = !append, quote_escape = "double") {
-  write_delim(x, path, delim = '\t', na = na, append = append, col_names = col_names, quote_escape = quote_escape)
+write_tsv <- function(x, path, na = "NA", append = FALSE, col_names = !append, quote_escape = "double", eol = "\n") {
+  write_delim(x, path, delim = '\t', na = na, append = append, col_names = col_names, quote_escape = quote_escape, eol = eol)
 }
 
 #' Convert a data frame to a delimited string
@@ -154,32 +156,32 @@ write_tsv <- function(x, path, na = "NA", append = FALSE, col_names = !append, q
 #' cat(format_csv(df))
 #' @export
 format_delim <- function(x, delim, na = "NA", append = FALSE,
-                         col_names = !append, quote_escape = "double") {
+                         col_names = !append, quote_escape = "double", eol = "\n") {
   stopifnot(is.data.frame(x))
 
   x[] <- lapply(x, output_column)
-  res <- stream_delim(df = x, path = NULL, delim = delim, col_names = col_names, append = append, na = na, quote_escape = quote_escape)
+  res <- stream_delim(df = x, path = NULL, delim = delim, col_names = col_names, append = append, na = na, quote_escape = quote_escape, eol = eol)
   Encoding(res) <- "UTF-8"
   res
 }
 
 #' @export
 #' @rdname format_delim
-format_csv <- function(x, na = "NA", append = FALSE, col_names = !append, quote_escape = "double") {
-  format_delim(x, delim = ",", na = na, append = append, col_names = col_names, quote_escape = quote_escape)
+format_csv <- function(x, na = "NA", append = FALSE, col_names = !append, quote_escape = "double", eol = "\n") {
+  format_delim(x, delim = ",", na = na, append = append, col_names = col_names, quote_escape = quote_escape, eol = eol)
 }
 
 #' @export
 #' @rdname format_delim
-format_csv2 <- function(x, na = "NA", append = FALSE, col_names = !append, quote_escape = "double") {
+format_csv2 <- function(x, na = "NA", append = FALSE, col_names = !append, quote_escape = "double", eol = "\n") {
   x <- change_decimal_separator(x, decimal_mark = ",")
-  format_delim(x, delim = ";", na = na, append = append, col_names = col_names, quote_escape = quote_escape)
+  format_delim(x, delim = ";", na = na, append = append, col_names = col_names, quote_escape = quote_escape, eol = eol)
 }
 
 #' @export
 #' @rdname format_delim
-format_tsv <- function(x, na = "NA", append = FALSE, col_names = !append, quote_escape = "double") {
-  format_delim(x, delim = "\t", na = na, append = append, col_names = col_names, quote_escape = quote_escape)
+format_tsv <- function(x, na = "NA", append = FALSE, col_names = !append, quote_escape = "double", eol = "\n") {
+  format_delim(x, delim = "\t", na = na, append = append, col_names = col_names, quote_escape = quote_escape, eol = eol)
 }
 
 #' Preprocess column for output
@@ -220,7 +222,7 @@ output_column.list <- function(x, name) {
   stop("Flat files can't store the list column `", name, "`", call. = FALSE)
 }
 
-stream_delim <- function(df, path, append = FALSE, bom = FALSE, ..., quote_escape) {
+stream_delim <- function(df, path, append = FALSE, bom = FALSE, ..., quote_escape, eol) {
   quote_escape <- standardise_escape(quote_escape)
 
   path <- standardise_path(path, input = FALSE)
@@ -233,7 +235,7 @@ stream_delim <- function(df, path, append = FALSE, bom = FALSE, ..., quote_escap
       open(path, "wb")
     }
   }
-  stream_delim_(df, path, ..., bom = bom, quote_escape = quote_escape)
+  stream_delim_(df, path, ..., bom = bom, quote_escape = quote_escape, eol = eol)
 }
 
 change_decimal_separator <- function(x, decimal_mark = ",") {