+++ src/RcppExports.cpp
@@ -352,8 +352,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // stream_delim_
-std::string stream_delim_(const List& df, RObject connection, char delim, const std::string& na, bool col_names, bool bom, int quote_escape);
-RcppExport SEXP _readr_stream_delim_(SEXP dfSEXP, SEXP connectionSEXP, SEXP delimSEXP, SEXP naSEXP, SEXP col_namesSEXP, SEXP bomSEXP, SEXP quote_escapeSEXP) {
+std::string stream_delim_(const List& df, RObject connection, char delim, const std::string& na, bool col_names, bool bom, int quote_escape, const char* eol);
+RcppExport SEXP _readr_stream_delim_(SEXP dfSEXP, SEXP connectionSEXP, SEXP delimSEXP, SEXP naSEXP, SEXP col_namesSEXP, SEXP bomSEXP, SEXP quote_escapeSEXP, SEXP eolSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -364,7 +364,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< bool >::type col_names(col_namesSEXP);
     Rcpp::traits::input_parameter< bool >::type bom(bomSEXP);
     Rcpp::traits::input_parameter< int >::type quote_escape(quote_escapeSEXP);
-    rcpp_result_gen = Rcpp::wrap(stream_delim_(df, connection, delim, na, col_names, bom, quote_escape));
+    Rcpp::traits::input_parameter< const char* >::type eol(eolSEXP);
+    rcpp_result_gen = Rcpp::wrap(stream_delim_(df, connection, delim, na, col_names, bom, quote_escape, eol));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -395,7 +396,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_readr_write_lines_raw_", (DL_FUNC) &_readr_write_lines_raw_, 3},
     {"_readr_write_file_", (DL_FUNC) &_readr_write_file_, 2},
     {"_readr_write_file_raw_", (DL_FUNC) &_readr_write_file_raw_, 2},
-    {"_readr_stream_delim_", (DL_FUNC) &_readr_stream_delim_, 7},
+    {"_readr_stream_delim_", (DL_FUNC) &_readr_stream_delim_, 8},
     {NULL, NULL, 0}
 };